fix init of barstyles to happen inside inspector
